# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Heyarr
  version: "1"
  summary: Headless content lifecycle, replication and consumption for homelabs.
  description: |
    Every error is an RFC 9457 problem document with a stable `type` URI. Clients
    branch on `type` — not on the title, not on the detail, and not on the HTTP
    status alone, which several different conditions share.

    Collections are paginated by keyset with an opaque cursor. `next_cursor` is
    absent on the last page, so "keep going while next_cursor is set" is the
    whole client loop. The cursor's contents are not a contract: it encodes the
    sort key of the last row returned, and a client that parses it breaks the
    first time a sort order changes.
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
servers:
  - url: http://127.0.0.1:8385
    description: The default loopback bind (ADR-0011).
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/renderers/{udn}/play:
    parameters:
      - name: udn
        in: path
        required: true
        description: The renderer's UPnP UDN, from `GET /api/v1/renderers`.
        schema: { type: string }
    post:
      tags: [playback]
      operationId: playOnRenderer
      summary: Plan a playback, mint a capability URL and push it to a renderer.
      description: >-
        One call, deliberately. Plan, mint and push always happen together, and
        a client made to orchestrate them is a client that can get the order
        wrong, hold a capability it never uses, or open a session for a
        playback it never started.


        The renderer's own `GetProtocolInfo` answer becomes the Device profile
        the planner decides against, so the file is judged on what that
        television said about itself rather than on what somebody typed in.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [asset_id]
              properties:
                asset_id: { type: string }
      responses:
        '200':
          description: Playing.
        '409':
          description: >-
            Nothing to play from — most often a replica on another peer, since
            a capability is only valid at the peer that signed it (ADR-0039).
        '502':
          description: >-
            The renderer refused, with its own UPnP error. `714 Illegal
            MIME-type` and `701 Transition not available` are different
            problems and need different fixes, so the device's answer is
            passed through rather than flattened.
...
  /api/v1/playback:
    post:
      tags: [playback]
      operationId: startPlayback
      summary: Plan, open a session, and return somewhere to play from.
      description: |
        The one call a client makes when someone presses play.

        **The controller stays out of the content data path (§32).** It returns
        a direct URL to the peer holding the replica; it does not proxy bytes.
        A controller that proxied would make playback's availability depend on
        its own, which is exactly the coupling §53's degraded-operation model
        exists to avoid — a Full Peer must keep streaming when the controller
        has gone.

        The URL is `/api/v1/blobs/{hash}/content`, unchanged. ADR-0013's
        consequences forbid "adding a player-shaped session token" to it, so
        playback gets no byte route, no auth scheme and no signed URL of its
        own — it gets a short-lived credential and the same endpoint
        replication and the BitTorrent web-seed use.

        **The credential is scoped to `read` and expires, but is not scoped to
        one blob.** A client holding one can read any blob until it expires.
        Per-blob scoping is a *grant* (§77's `/grants`), which is a later
        milestone with principals behind it; inventing a lesser version here
        would be a second authorisation model to reconcile with the real one.

        **Anything that is not DIRECT is a 409** carrying the planner's first
        reason, not a 501. The request is well-formed and the asset is real;
        what cannot happen is this pairing of asset and device. No session is
        opened — a session for a playback that cannot happen is state nobody
        will clean up, and it would appear in "continue watching" for something
        nobody watched. `POST /playback/plan` has the full rationale.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StartPlayback" }
      responses:
        "201":
          description: The session, the plan, and where to play from.
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StartedPlayback" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: |
            The plan is not DIRECT, or nothing holds the bytes. No session was
            opened.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
...
  /api/v1/consumption/sessions:
    get:
      tags: [playback]
      operationId: listConsumptionSessions
      summary: Page consumption sessions, newest first.
      description: |
        `?state=resumable` is the **"continue watching"** query: every session
        that has not reached a terminal state, most recently touched first, for
        watching, listening and reading alike (§67).

        It is a named filter rather than `state=created,playing,paused` because
        "resumable" is the question clients actually ask, and spelling it as
        three states would put the state machine's shape into every client.
      parameters:
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
        - name: state
          in: query
          schema:
            type: string
            enum: [resumable, created, playing, paused, stopped, completed]
        - name: asset_id
          in: query
          schema: { type: string }
        - name: device_id
          in: query
          schema: { type: string }
      responses:
        "200":
          description: A page of sessions.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/ConsumptionSession" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [playback]
      operationId: createConsumptionSession
      summary: Open a consumption session.
      description: |
        One model covers watching, listening and reading (§67, ADR-0024). The
        session starts in `created` — a real state, not a formality: the gap
        between "I intend to watch this" and "bytes are moving" is where the
        planner runs, and collapsing it makes an abandoned plan
        indistinguishable from a playback that stalled instantly.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateConsumptionSession" }
      responses:
        "201":
          description: The session.
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConsumptionSession" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        An opaque scoped token (ADR-0011), presented as `Authorization: Bearer
        heyarr_…`. The header is the only place a credential is accepted: a
        `?token=` parameter would write it into every proxy log, browser history
        and Referer header between the client and here.

        Scopes are ordered — `admin` implies `write` implies `read`. Every
        route under /api/v1 requires at least `read`.

    peerCertificate:
      type: mutualTLS
      description: |
        A self-signed X.509 certificate carrying this peer's Ed25519 public
        key, pinned against the controller-issued membership record
        (ADR-0012). There is no CA and no PKI: the certificate proves
        possession of a key, and membership decides whether that key may
        speak. Both ends pin, so the dialler authenticates the listener by
        exactly the same rule.

        It is checked twice — during the handshake, and again on every request
        the connection carries. Revocation is deleting the membership record,
        and there is no revocation list to fall back on, so the freshness of
        that second lookup is the mechanism.

        This credential is accepted ONLY on /peer/v1, and a bearer token is
        never accepted there. A connection presenting no client certificate is
        refused during the handshake, not with a status code.
  parameters:
...
    cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: |
        An opaque position from a previous page's `next_cursor`. It is scoped to
        one collection: a cursor from `/works` is rejected by `/jobs` rather
        than being read as a position in a different sort order.
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      description: Values above the maximum are clamped, not rejected.
  responses:
    BadRequest:
      description: The request is malformed or semantically invalid.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: No credential, or one that was rejected.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: A valid credential that does not carry the required scope.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: |
        No such resource. Never an empty 200: a client cannot tell "no such
        work" from "a work with nothing in it" if both are 200, and it is the
        difference between retrying and giving up.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
    Problem:
      type: object
      description: RFC 9457. Clients branch on `type` and nothing else.
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        request_id:
          type: string
          description: Correlates the response with the server log line.
    Page:
      type: object
      required: [items]
      properties:
        items: { type: array, items: {} }
        next_cursor:
          type: string
          description: Absent on the last page.
...
    StartPlayback:
      type: object
      required: [asset_id, device_id]
      properties:
        asset_id: { type: string }
        device_id: { type: string }
        verb:
          type: string
          enum: [watch, listen, read]
          description: |
            Optional. Derived from the probe when omitted — watching when there
            is video, listening when there is not. Derived from the probe
            rather than the extension because a `.mkv` holding only audio is a
            legitimate thing, and calling it "watching" puts it in the wrong
            row of every continue-watching list.
    StartedPlayback:
      type: object
      properties:
        session_id: { type: string }
        plan: { $ref: "#/components/schemas/PlaybackPlan" }
        content_url:
          type: string
          description: |
            The ordinary blob endpoint (ADR-0013), on the peer routing selected
            (§32) — **never the controller**, which stays out of the content
            data path. Absolute when the source is another peer, so the client
            connects to that peer directly; relative when the source is this
            node, which is the origin the client is already talking to
            (ADR-0029).

            Bounded by the credential below, which expires. It is deliberately
            NOT scoped to one blob: per-blob scoping is a grant (§77) and §54's
            lease shape is the real answer, arriving in M7 alongside degraded
            authorisation. Revisit when the first peer class arrives whose
            access is a strict subset of the canonical set (§6, §19) — the same
            trigger ADR-0030 records.
        token:
          type: string
          description: |
            A short-lived `read` credential for that URL. **Not scoped to one
            blob** — see the endpoint description.
        expires_at: { type: string, format: date-time }
...
    PlaybackPlan:
      type: object
      properties:
        asset_id: { type: string }
        device_id: { type: string }
        decision:
          type: string
          enum: [direct, remux, transcode, unplayable]
        reasons:
          type: array
          description: |
            Why the decision is not DIRECT, or why nothing can be played. Empty
            for a clean DIRECT.
          items: { $ref: "#/components/schemas/PlanReason" }
        peer_id:
          type: string
          description: The replica chosen to serve from. Absent when unplayable.
        remote:
          type: boolean
          description: |
            The chosen replica is at another site. §31 prefers a local one, so
            an instance where this is usually true is one where replication is
            not working.

            **The unproven note this carried is retired (M4-11, M4-14).** It
            said nothing had ever run against a second peer; one exists now, and
            §32's source selection routes against it — the same client with and
            without a local replica is routed two different ways, and the
            cross-site answer records that it fell back.

            Read `routing` rather than this flag when you want to know why. This
            is the planner's own summary of the source it was handed; `routing`
            is the decision, with the peers it rejected and the reason for each.
        content_url:
          type: string
          description: |
            Where the bytes are, for a DIRECT plan — on the peer `routing`
            selected, not on the controller. **Absent otherwise** — a REMUX or
            TRANSCODE plan has no bytes to point at yet, and offering the
            original would invite the client to play exactly what the plan just
            said it cannot.
        routing:
          allOf: [{ $ref: "#/components/schemas/RoutingDecision" }]
          description: |
            Which peer this read was routed to, and why not the others (§32).
            Present on every plan **including a refusal** — especially on a
            refusal. "Unavailable" with nothing after it cannot tell a peer that
            is down from a peer that never had the bytes, and those have
            different fixes.
...
    PlanReason:
      type: object
      properties:
        code:
          type: string
          enum:
            [no_replica, container_unsupported, video_codec_unsupported,
             audio_codec_unsupported, resolution_too_high, bitrate_too_high,
             hdr_unsupported, no_probe, remote_replica_only,
             device_declares_nothing]
          description: Stable. Clients branch on this, and a client branching on prose breaks when the prose improves.
        detail:
          type: string
          description: Prose for a human. May change.
    ConsumptionSession:
      type: object
      description: |
        One consumption of one asset by one device — watching, listening or
        reading (§67, ADR-0024).
      properties:
        id: { type: string }
        asset_id: { type: string }
        device_id: { type: string }
        verb:
          type: string
          enum: [watch, listen, read]
        state:
          type: string
          enum: [created, playing, paused, stopped, completed]
        progress:
          allOf:
            - $ref: "#/components/schemas/SessionProgress"
          description: |
            **Absent** until something has been recorded — not a zero locator.
            Position zero is a real position, and a client cannot tell "at the
            very beginning" from "never played" if both render as 0.
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        started_at:
          type: string
          format: date-time
          nullable: true
          description: |
            When the session first played. Null for one created and abandoned,
            which is what makes "nobody watched this" distinguishable from
            "someone watched two minutes".
        ended_at:
          type: string
          format: date-time
          nullable: true
          description: When it reached a terminal state.
    SessionProgress:
      type: object
      description: |
        Where a session reached, as a **locator and a unit** — never a bare
        number. A page number is not a number of seconds, and storing both as
        "position" makes every reader guess which it has.
      properties:
        locator:
          type: string
          description: |
            A decimal offset for `seconds`, a decimal page number for `page`,
            or an `epubcfi(...)` expression for `cfi`. A CFI is opaque to
            Heyarr, which stores and serves publications and never renders
            them (§69).
        unit:
          type: string
          enum: [seconds, page, cfi]
    CreateConsumptionSession:
      type: object
      required: [asset_id, device_id, verb]
      properties:
        asset_id: { type: string }
        device_id: { type: string }
        verb:
          type: string
          enum: [watch, listen, read]
...